ml/requirements.txt
Flask==3.0.3
gunicorn>=20.1.0
python-dotenv==0.21.0
requests==2.32.3
scikit-image==0.24.0
numpy==1.26.4
matplotlib==
boto3
python-decouple
flask_cors